fix: JPQL SUM 집계 반환 타입 수정

+++ src/main/java/navik/domain/kpi/dto/res/KpiScoreResponseDTO.java
@@ -39,8 +39,8 @@ public class KpiScoreResponseDTO {
 	public record MonthlyTotalScoreChange(
 		int year,
 		int month,
-		int currentScore,
-		int previousScore,
+		long currentScore,
+		long previousScore,
 		Double changeRate
 	) {
 	}
